Add selecting methods to NodifyEditor
