+++ Nodify/EditorStates/EditorSelectingState.cs
@@ -1,5 +1,4 @@
 ﻿using System.Windows.Input;
-using static Nodify.SelectionHelper;
 
 namespace Nodify
 {
@@ -7,45 +6,41 @@ namespace Nodify
     public class EditorSelectingState : EditorState
     {
         private readonly SelectionType _type;
-        private bool _canceled;
-
-        /// <summary>The selection helper.</summary>
-        protected SelectionHelper Selection { get; }
+        private bool Canceled { get; set; } = NodifyEditor.AllowSelectionCancellation;
 
         /// <summary>Constructs an instance of the <see cref="EditorSelectingState"/> state.</summary>
         /// <param name="editor">The owner of the state.</param>
         /// <param name="type">The selection strategy.</param>
         public EditorSelectingState(NodifyEditor editor, SelectionType type) : base(editor)
         {
-            Selection = new SelectionHelper(editor);
             _type = type;
         }
 
         /// <inheritdoc />
         public override void Enter(EditorState? from)
         {
-            Editor.UnselectAllConnection();
+            Canceled = false;
 
-            _canceled = false;
-            Selection.Start(Editor.MouseLocation, _type);
+            Editor.BeginSelection(Editor.MouseLocation, _type);
         }
 
         /// <inheritdoc />
         public override void Exit()
         {
-            if (_canceled)
+            // TODO: This is not canceled on LostMouseCapture (add OnLostMouseCapture/OnCancel callback?)
+            if (Canceled)
             {
-                Selection.Abort();
+                Editor.CancelSelection();
             }
             else
             {
-                Selection.End();
+                Editor.EndSelection();
             }
         }
 
         /// <inheritdoc />
-        public override void HandleMouseMove(MouseEventArgs e)
-            => Selection.Update(Editor.MouseLocation);
+        public override void HandleMouseMove(MouseEventArgs e) 
+            => Editor.UpdateSelection(Editor.MouseLocation);
 
         /// <inheritdoc />
         public override void HandleMouseDown(MouseButtonEventArgs e)
@@ -60,12 +55,15 @@ namespace Nodify
         public override void HandleMouseUp(MouseButtonEventArgs e)
         {
             EditorGestures.SelectionGestures gestures = EditorGestures.Mappings.Editor.Selection;
-
-            bool canCancel = gestures.Cancel.Matches(e.Source, e);
-            bool canComplete = gestures.Select.Matches(e.Source, e);
-            if (canCancel || canComplete)
+            if(gestures.Select.Matches(e.Source, e))
             {
-                _canceled = !canComplete && canCancel;
+                PopState();
+            }
+            else if(NodifyEditor.AllowSelectionCancellation && gestures.Cancel.Matches(e.Source, e))
+            {
+                Canceled = true;
+                e.Handled = true;   // prevents opening context menu
+
                 PopState();
             }
         }
@@ -76,9 +74,10 @@ namespace Nodify
 
         public override void HandleKeyUp(KeyEventArgs e)
         {
-            if (EditorGestures.Mappings.Editor.Selection.Cancel.Matches(e.Source, e))
+            EditorGestures.SelectionGestures gestures = EditorGestures.Mappings.Editor.Selection;
+            if (NodifyEditor.AllowSelectionCancellation && gestures.Cancel.Matches(e.Source, e))
             {
-                _canceled = true;
+                Canceled = true;
                 PopState();
             }
         }
